nexus/backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.6.1
SQLAlchemy==2.0.36
python-dotenv==1.0.1
pandas==2.2.3
numpy==2.1.3
requests==2.32.3
passlib[bcrypt]==1.7.4
python-jose==3.3.0
